[NRT-766] Extract generic RichTooltip from sidebar tooltip
The interactive tooltip (clickable links + stays open on hover) isn't
sidebar-specific. Split it into a reusable RichTooltip in rich_tooltip.py
(owns the popup, wake-up delay, stay-open polling, and native styling, driven
by a target_at(global_pos) -> (html, anchor_rect) callback); sidebar_tooltip.py
is now a thin adapter mapping a cursor position to the hovered item's HTML and
row rect. Ready to reuse for the deferred disabled-button / context-menu
tooltips. No behavior change.
